pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "podsignal"
version = "1.1.0"
description = "Podcast guest intelligence pipeline for B2B lead generation"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.25.0",
    "requests>=2.31.0",
    "feedparser>=6.0.10",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
]

[project.scripts]
podsignal = "podsignal.__main__:main"

[tool.setuptools.packages.find]
include = ["podsignal*"]
